Add files via upload
Our website provides a start front page of index where you can start and go to sign up page where you can sign up and make a profile as a user and as a mentor both ,after sign in you will go to home page at which there are pop up motivational and calming quotes are there and also three options of games ,videos and menu are there. By clickig on menu a page of selecting mental disorders which you have or might have myth of, After that there will be a quiz based on that and after attempting that quiz a score card will shown if your score will be less than the decided ,then you will be given selfcare tips like meditation exercices,etc and if it is above the average score then a chat box will open in which you can directly consult with mentors . In videos button , calm , relaxing and motivated videos like insta reels will be shown and in games you can play stress relieving games like popping the balloon and get score also of that and also a puzzle game which will divert your mind and also good for your brain health.